restructure slightly and fix unused imports + add todo for CLI
