+++ src/bin/sqlite_server_cli.rs
@@ -1,204 +1,10 @@
-use core::str;
-use std::fs;
-use std::path::Path;
-use std::{env, path::PathBuf};
+use std::env;
 
-use regex::Regex;
-use sha2::{Digest, Sha256};
-use sqlite_server::core::{
-    constants::queries,
-    db::{execute_query, AppliedQuery, QueryArg},
-    util::get_flag_val,
-};
-use sqlx::{migrate::MigrateDatabase, Sqlite, SqlitePool};
+use sqlite_server::cli::util::{get_flag_val, DatabaseManager};
+use sqlite_server::core::constants::cli;
+use sqlx::{migrate::MigrateDatabase, Sqlite};
 
-include!(concat!(env!("OUT_DIR"), "/gen.rs"));
-
-struct DatabaseManager {
-    pub consumer_db_base_path: PathBuf,
-    pub user_db_base_path: PathBuf,
-    pub user_db_full_path_string: String,
-}
-
-impl DatabaseManager {
-    pub fn new() -> DatabaseManager {
-        let root_dir = Path::new(ROOT_DIR);
-        let cfg_path = root_dir.join("cfg");
-        let consumer_db_base_path = root_dir.join("db");
-        let user_db_base_path = cfg_path.join(USER_DB_HASH);
-        let user_db_full_path_string = format!(
-            "{}/{}.db",
-            user_db_base_path.to_str().unwrap(),
-            USER_DB_HASH
-        );
-
-        DatabaseManager {
-            consumer_db_base_path: consumer_db_base_path,
-            user_db_base_path: user_db_base_path,
-            user_db_full_path_string: user_db_full_path_string,
-        }
-    }
-
-    pub async fn init(&self) {
-        if !self.user_db_base_path.exists() {
-            let _ = fs::create_dir_all(&self.user_db_base_path);
-        }
-
-        if !self.consumer_db_base_path.exists() {
-            let _ = fs::create_dir_all(&self.consumer_db_base_path);
-        }
-
-        match Sqlite::create_database(&self.user_db_full_path_string).await {
-            Ok(_) => {
-                let pool =
-                    SqlitePool::connect(&format!("sqlite:{}", self.user_db_full_path_string))
-                        .await
-                        .unwrap();
-                let mut transaction = pool.begin().await.unwrap();
-
-                // CREATE users
-                match execute_query(
-                    AppliedQuery::new(queries::CREATE_USERS_TABLE),
-                    &mut *transaction,
-                )
-                .await
-                {
-                    Ok(_) => {
-                        // CREATE database accesses
-                        // access_right can contain, or at least handle 1, 2 or 3 as value
-                        // all other values will be seen as non-functioning
-                        match execute_query(
-                            AppliedQuery::new(queries::CREATE_USERS_DATABASE_ACCESS_TABLE),
-                            &mut *transaction,
-                        )
-                        .await
-                        {
-                            Ok(_) => {
-                                let _ = transaction.commit().await;
-                            }
-                            Err(err) => {
-                                let _ = transaction.rollback().await;
-                                panic!("Error: {}", err);
-                            }
-                        }
-                    }
-                    Err(err) => {
-                        let _ = transaction.rollback().await;
-                        panic!("Error: {}", err);
-                    }
-                }
-            }
-            Err(err) => panic!("Error: {}", err),
-        }
-    }
-
-    pub async fn create_database(&self, db_name: &str) {
-        if !db_name.eq("") {
-            let regex = Regex::new(r"^[a-zA-Z0-9_-]+$").unwrap();
-            if !regex.is_match(db_name) {
-                panic!("Error: Database name format must follow either one or a combination of the patterns [a-z, A-Z, 0-9, _, -]");
-            }
-
-            let db_name_hash = base16ct::lower::encode_string(&Sha256::digest(db_name.as_bytes()));
-            let consumer_db_full_path_string = format!(
-                "{}/{}",
-                self.consumer_db_base_path.to_str().unwrap(),
-                db_name_hash
-            );
-            let consumer_db_full_path = Path::new(&consumer_db_full_path_string);
-            let consumer_db = format!("{}/{}.db", consumer_db_full_path_string, db_name_hash);
-
-            if !consumer_db_full_path.exists() {
-                let _ = fs::create_dir_all(&consumer_db_full_path);
-            }
-
-            if !Sqlite::database_exists(&consumer_db).await.unwrap_or(false) {
-                match Sqlite::create_database(&consumer_db).await {
-                    Ok(_) => {
-                        println!("Successfully created db {} as {}", db_name, db_name_hash);
-                        let _ = fs::write(
-                            format!("{}/{}", consumer_db_full_path_string, db_name),
-                            db_name_hash,
-                        );
-                    }
-                    Err(err) => panic!("Error: {}", err),
-                }
-            } else {
-                panic!("Error: Database already exists");
-            }
-        }
-    }
-
-    pub async fn create_user(&self, username: &str, password: &str) {
-        if !username.eq("") && !password.eq("") {
-            let username_hash =
-                base16ct::lower::encode_string(&Sha256::digest(username.as_bytes()));
-            let username_password_hash = base16ct::lower::encode_string(&Sha256::digest(
-                format!("{}{}", username, password).as_bytes(),
-            ));
-
-            let pool = SqlitePool::connect(&format!("sqlite:{}", self.user_db_full_path_string))
-                .await
-                .unwrap();
-            let mut transaction = pool.begin().await.unwrap();
-
-            let _ = match execute_query(
-                AppliedQuery::new(queries::INSERT_USER).with_args(vec![
-                    QueryArg::String(username),
-                    QueryArg::String(&username_hash),
-                    QueryArg::String(&username_password_hash),
-                ]),
-                &mut *transaction,
-            )
-            .await
-            {
-                Ok(_) => transaction.commit().await,
-                Err(err) => {
-                    let _ = transaction.rollback().await;
-                    panic!("Error: {}", err);
-                }
-            };
-        } else {
-            panic!("Error: Must provide username and password with flags [-u, -p]");
-        }
-    }
-
-    pub async fn modify_user_access(&self, username: &str, database_name: &str, access_right: i32) {
-        if !username.eq("") && !database_name.eq("") && access_right != 0 {
-            let username_hash =
-                base16ct::lower::encode_string(&Sha256::digest(username.as_bytes()));
-            let database_name_hash =
-                base16ct::lower::encode_string(&Sha256::digest(database_name.as_bytes()));
-
-            let pool = SqlitePool::connect(&format!("sqlite:{}", self.user_db_full_path_string))
-                .await
-                .unwrap();
-            let mut transaction = pool.begin().await.unwrap();
-
-            let _ = match execute_query(
-                AppliedQuery::new(queries::UPSERT_USER_DATABASE_ACCESS).with_args(vec![
-                    QueryArg::String(database_name),
-                    QueryArg::String(&database_name_hash),
-                    QueryArg::Int(access_right),
-                    QueryArg::String(&username_hash),
-                ]),
-                &mut *transaction,
-            )
-            .await
-            {
-                Ok(_) => {
-                    let _ = transaction.commit().await;
-                }
-                Err(err) => {
-                    let _ = transaction.rollback().await;
-                    panic!("Error: {}", err);
-                }
-            };
-        } else {
-            panic!("Error: Must provide username, database and access right(1-3) with flags [-u, -db, -a]");
-        }
-    }
-}
+// TODO: ADD proper struct for CLI args management
 
 // DEFAULTS:
 //  (root path) $HOME/.serf/
@@ -215,7 +21,8 @@ async fn main() -> std::io::Result<()> {
         .await
         .unwrap_or(false)
     {
-        database_manager.init().await
+        database_manager.init().await;
+        println!("INITIAL SETUP");
     }
 
     let args: Vec<String> = env::args().collect();
@@ -229,12 +36,14 @@ async fn main() -> std::io::Result<()> {
 
                 match cmd_two {
                     "database" => {
-                        let db = get_flag_val::<String>(&args_split, "-db").unwrap();
+                        let db = get_flag_val::<String>(&args_split, cli::DB_NAME_FLAG).unwrap();
                         database_manager.create_database(&db).await;
                     }
                     "user" => {
-                        let username = get_flag_val::<String>(&args_split, "-u").unwrap();
-                        let password = get_flag_val::<String>(&args_split, "-p").unwrap();
+                        let username =
+                            get_flag_val::<String>(&args_split, cli::USERNAME_FLAG).unwrap();
+                        let password =
+                            get_flag_val::<String>(&args_split, cli::PASSWORD_FLAG).unwrap();
                         database_manager.create_user(&username, &password).await;
                     }
                     _ => panic!(
@@ -250,9 +59,12 @@ async fn main() -> std::io::Result<()> {
 
                     match cmd_three {
                         "access" => {
-                            let username = get_flag_val::<String>(&args_split, "-u").unwrap();
-                            let database = get_flag_val::<String>(&args_split, "-db").unwrap();
-                            let access_right = get_flag_val::<i32>(&args_split, "-a").unwrap();
+                            let username =
+                                get_flag_val::<String>(&args_split, cli::USERNAME_FLAG).unwrap();
+                            let database =
+                                get_flag_val::<String>(&args_split, cli::DB_NAME_FLAG).unwrap();
+                            let access_right =
+                                get_flag_val::<i32>(&args_split, cli::ACCESS_RIGHT_FLAG).unwrap();
 
                             database_manager
                                 .modify_user_access(&username, &database, access_right)
