+++ src/core/constants/cli.rs
@@ -4,4 +4,8 @@ pub const DEFAULT_DB_MAX_IDLE_TIME: u64 = 3600;
 pub const PORT_FLAG: &str = "--port";
 pub const DB_MAX_CONN_FLAG: &str = "--db-max-conn";
 pub const DB_MAX_IDLE_TIME_FLAG: &str = "--db-max-idle-time";
+pub const DB_NAME_FLAG: &str = "-db";
+pub const USERNAME_FLAG: &str = "-u";
+pub const PASSWORD_FLAG: &str = "-p";
+pub const ACCESS_RIGHT_FLAG: &str = "-ar";
