+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod core;
 pub mod web;
+pub mod cli;
