+++ src/web/mod.rs
@@ -1,3 +1,4 @@
 pub mod controller;
 pub mod jwt;
 pub mod request;
+pub mod util;
